+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-	"editor.formatOnSave": true
+  "editor.formatOnSave": true
 }
