+++ package.json
@@ -16,10 +16,13 @@
 		"katex": "^0.16.3",
 		"next": "12.3.1",
 		"react": "^17.0.2",
+		"react-dnd": "^16.0.1",
+		"react-dnd-html5-backend": "^16.0.1",
 		"react-dom": "^17.0.2",
 		"react-icons": "^4.7.1",
 		"react-markdown": "^8.0.3",
 		"rehype-katex": "^6.0.2",
+		"rehype-raw": "^6.1.1",
 		"remark-gfm": "^3.0.1",
 		"remark-math": "^5.1.1"
 	},
