+++ src/components/Courses/Material/Components/Graph/index.ts
@@ -1,0 +1,1 @@
+export * from "./CourseMaterialGraph";
