+++ src/components/Courses/Material/Components/Match/index.ts
@@ -1,0 +1,3 @@
+export * from "./CourseMaterialMatchBox";
+export * from "./CourseMaterialMatchCard";
+export * from "./CourseMaterialMatchDrop";
