+++ src/components/Courses/Material/Components/index.ts
@@ -1,0 +1,2 @@
+export * from "./Graph";
+export * from "./Match";
