+++ src/components/Courses/Material/index.ts
@@ -1,0 +1,3 @@
+export * from "./CourseMaterialContent";
+export * from "./Components";
+export * from "./Side";
