--- src/components/basic/Loading/index.ts
@@ -1,1 +1,0 @@
-export * from "./Loader";
