+++ src/components/basic/index.ts
@@ -1,7 +1,7 @@
 export * from "./Badge";
 export * from "./Button";
 export * from "./Input";
+export * from "./Loader";
 export * from "./Quote";
 export * from "./Section";
 export * from "./ToggleLink";
-export * from "./Loading";
