+++ src/components/courses/index.ts
@@ -1,1 +1,2 @@
+export * from "./Material";
 export * from "./ProgressVertical";
