--- src/components/materials/Match/index.ts
@@ -1,2 +1,0 @@
-export * from "./MatchBox";
-export * from "./MatchCard";
