--- src/components/materials/index.ts
@@ -1,3 +1,0 @@
-export * from "./Content";
-export * from "./Graph";
-export * from "./Side";
