+++ src/contexts/SwapPageContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+import { StateType } from "../type";
+
+export const SwapPageContext = createContext<StateType<boolean>>([
+  false,
+  () => {},
+]);
