+++ src/hooks/materials/useSwapPage.ts
@@ -1,0 +1,9 @@
+import { SwapPageContext } from "@/src/contexts";
+import { useContext, useMemo } from "react";
+
+export function useSwapPage() {
+  const stateSwapPage = useContext(SwapPageContext);
+  console.log("Something Changed!");
+  console.log(stateSwapPage[0]);
+  return useMemo(() => stateSwapPage, [stateSwapPage]);
+}
