+++ src/materials/calculus/derivative/definition.mdx
@@ -1,3 +1,5 @@
+import
+
 # Definitions
 
 ## Precise Definition of Limit
@@ -12,7 +14,7 @@ Let $f$ be a function defined on an open interval containing $a$, with the possi
 
 # Determining Limits
 
-Let's use a simple function $f(x) = x^2$ below. `[graph]function:x*x`
+Let's use a simple function $f(x) = x^2$ below. <Graph functions="x^2" />
 
 If we set $x = -1$, we will have $f(-1)=1$. If we set $x = -0.1$, we will have $f(-0.1)=0.01$. If we set $x = -0.01$, we will have $f(-0.01)=0.0001$.
 
@@ -24,7 +26,7 @@ In this case, we can say that the limit of $f(x)$ as $x$ approaches $0$, is $0$.
 
 # Determining Limits
 
-`[graph]function:1/x`
+<Graph functions="1/x" />
 
 Sometimes, we may get something like the above $g(x) = \frac{1}{x}$. What would be the value of $\lim\limits_{x\rightarrow 0} \frac{1}{x}$? That's right, it doesn't exist. If we recall the definiton of limit, we can clearly see that if we try approaching $x = 0$ from the right side, $g(x)$ will eventually approach $-\infty$, whereas approaching $x = 0$ from the left side, $g(x)$ will eventually approach $\infty$. Therefore, it is true that $\lim\limits_{x\rightarrow 0^-} \frac{1}{x} = -\infty$ and $\lim\limits_{x\rightarrow 0^+} \frac{1}{x} = \infty$.
 
@@ -32,7 +34,7 @@ Sometimes, we may get something like the above $g(x) = \frac{1}{x}$. What would 
 
 # Determining Limits
 
-`[graph]function:1/[x]`
+<Graph functions="1/[x]" />
 
 This case is a little bit deifferent than the last one. This time, we have $h(x) = \frac{1}{|x|}$ . As you may have guessed, this one function does have a limit as $x$ approaches $0$ because the limit as $x$ approaches $0$ from the right and from the left is the same. Therefore, we can say that $\lim\limits_{x\rightarrow 0} \frac{1}{|x|} = \infty$
