+++ src/materials/calculus/limits/common.mdx
@@ -26,7 +26,7 @@ However, we will try to **prove** it, and there are many ways to prove this limi
 
 # Proof of a Trigonometric Limit
 
-> #formula#
+> #theorem#
 > **Squeeze Theorem**
 >
 > Suppose that $f(x) ≤ g(x) ≤ h(x)$ for all $x$ in an open interval containing $a$, except possibly at $a$, and $\lim\limits_{x\rightarrow a}f(x)= \lim\limits_{x\rightarrow a} h(x)=L$. So, $\lim\limits_{x\rightarrow a} g(x)=L$.
@@ -81,7 +81,7 @@ $$
 $$
 
 There we go! We have proven that $\lim\limits_{x\rightarrow 0}\frac{\sin{x}}{x} = 1$.
-``
+
 ===
 
 # Limits of Logarithmic-related Functions
@@ -105,7 +105,7 @@ Since $x$ is approaching $\infty$, when we swap out the variables of the limit, 
 
 ---
 
-> #formula# Given the constant $a$, $\lim\limits\_{x\rightarrow 0} \frac{a^x - 1}{x} = \ln{a} $
+> #formula# Given the constant $a$, $\lim\limits_{x\rightarrow 0} \frac{a^x - 1}{x} = \ln{a} $
 
 To prove the above limit, we must first know that $a^x$ can be represented as a series, that is:  
 $$1 + x \ln{a} + \frac{\left ( x \ln{a} \right )^2}{2!} + \frac{\left ( x \ln{a} \right )^3}{3!} + ...$$
