+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./capitalize";
 export * from "./contexts";
 export * from "./course";
 export * from "./evaluateMath";
+export * from "./materials";
